Remove the pid argument from container_{g,p}_add_pid().
Removal of this argument forces the use of this function call only for
adding the calling process - and not any other process - to the job
container. This is needed for an in-development job container
plugin that adds jobs to namespaces. As it uses setns() system call -
which can only work on the calling process - the plugin thus relies
on being called from the calling process.

Bug 6567.