Fix for het jobs ability to run steps
Without this change a heterogeneous job allocation could not run
a step on any single component of a hetjob, but only across multiple
hetjob components. Note that by default a step runs on only the
first hetjob component, so the default mode of operation running
steps on a hetjob was broken. This bug caused tests 38.7 and 38.11
to fail consistently. The bug was introduced in commit 2f162b7a7addf