Store reservation flags in slurmdbd in a uint64_t.
Add warning to slurm.h.in that no new reservation flags can be
stored in slurmdbd in 19.05. (Although they could still be used by
slurmctld without issue.)

Note that the underlying RPC still uses uint32_t, but this will be
changed before 20.02 on master, and changing the column to uint32_t
in 19.05 just to change it again in 20.02 is best avoided.

Bug 6969.