Add _pack_list_of_str() and convert to it. No functional change.
This started as cleanup to use _list_count_null(), but once I realized
the amount of duplicated code here I decided to add this new static
function instead. This function, or one similar, should probably be moved
to pack.h instead.

Note that this cannot use slurm_pack_list(), as packstr() is a macro
not a function.