Use slurm_pack_list(). RPC format unchanged.
This does constitute a minor functional change - slurm_pack_list()
will insert NO_VAL as the count and rewind the buffer instead of
packing the count and as much can fit if the buffer runs out of space.

Given that the current code has no provision for error handling on the
pack side, I do not believe the addition of this error handling should
pose a problem.

Remove _list_count_null() now that it is unused.