Use slurm_pack_list() for sending accounting_list.
There is a subtle difference here. The prior code would send NO_VAL for the
count if the list did not exist, and if there were no elements in the list.

slurm_pack_list() will send a count of 0 in the later case, and create the
empty list for the remote to use.

After looking through the code that works with the accounting_list field,
both of these cases appear to be processed correctly, and the minor change
in behavior here should not cause any problems.