Use _pack_list_of_str() in slurmdb_pack_assoc_rec().
There is a subtle change hiding in a prior commit here. Note that
previously the count was not reset to NO_VAL after use for packing
other lists, which means that if qos_list did not exist the
accounting_list count was packed erroneously, which may have lead
to an unexpected unpack error on the receiving side.