Enhance test with job name and mpi option
This change adds a job name to all tests spawned by the test.
It also explicitly sets the MPI type to none. This is required
by some of the tests if using OpenMPI in multi-slurmd mode.
See note in test1.88 for full description of OpenMPI limitations
in this Slurm mode.