optimization of gres node filtering logic
This filters out usable nodes earlier in the logic. The previous
logic would filter out nodes early if the job specified a value
for gres-per-node (ie. --gres), but this extents the logic to
filter out nodes early when the gres-per-socket or gres-per-task
are specified using the count which is higher. For example, if
a job lacks a gres-per-node specification, but gres-per-task=2
then we can filter out nodes with less than 2 gres at this early
stage in the scheduling.

Bug 7085