Change test to better match sacct behavior
The sacct command in verison 19.05 when job ID is specified will
find all examples of that job ID run at any time. That means if
the job IDs numbers wrap around, this test will always fail. This
adds a start time to the sacct command of 00:00 (midnight of current
day) to avoid problems with wrapping job IDs and make this test
work more like it did in version 18.08. Note this test does have
a very tiny window for failures if the test program ran just before
midnight and the sacct command to view it's state ran just after
midnight. Given that the entire test only runs for a minute, that
is unlikely in practice.