Add deadlock detection to _mysql_query_internal().
Detected when there is a deadlock and retry the query up to
MAX_DEADLOCK_ATTEMPTS. If it doesn't work, then there is
something else wrong and give up. This will most likely
cause another connection to MySQL. Also instruct the user to
open a ticket as there may be a more structural issue with
the query than a race condition with another parallel query.

Bug 7161.