add suffix support between fields in sinfo/squeue long format
This change adds the ability to add suffix value between fields in the
long formats of sinfo and squeue so that we can duplicate the behavior
of the short options. For example, there was no way to replicate something
like "%P %V" in squeue as something like "partition:9,cluster:8" would
just run the two fields together without a space and something like
"partition:9 ,cluster:8" would generate an error. This change treats
anything between the number and comma as a suffix to be included in
the output and make this work.

bug 7186