Modify sinfo/squeue --noheader format for multiple clusters
The default output for the sinfo and squeue commands when using
the --noheader format removes the "CLUSTER: name" header for each
block and adds the cluster name to each line in the default output
format option. If the user has explicitly specified the output
format, then the user's format specification will be used.

bug 7186