Fix (un)supported states in sacct
Add explicit support for BOOT_FAIL.
Rearrange db queries for "finished" states assuming -S and -E are
correctly set, as with the other states.
Add error() message to dbd if unsupported state is requested.
Add _supported_state() to the client and fatal() if unsupported state
is requested by the user of the client.

Bug 7110