Fix unaccounted TRESRunMins usage from HetJobs
The implementation of priority_p_job_end in priority/multifactor
expects the job state to be set to complete or completing in order to
properly remove some job usage from the assoc and qos. This must be
simulated by the pack job run check code, or the check-time usage is not
removed.

Bug 7284