Use ftruncate() instead of lseek() + write().
Functionally equivalent.

This does discard the potential call to stepd_drain_node() on ENOSPC
from the write(). However in testing, most filesystems will create
sparse files with this sequence of calls, and the write will not actually
use up the full space the file needs, and thus never returns ENOSPC here
even if the filesystem is nearly full.