Allow account coordinators to add new users.
Account coordinators are allowed to add users to their accounts, but
cannot change the default account of a user. However, when a user is
added to an account and the user doesn't have an account yet, the
account is set as the user's default account, which counts as changing
the user's default account. This commit allows account coordinators to
add users who don't already have a default account elsewhere to the
accounts they are coordinators over.

Bug 7018.