Docs - fix regex used to pick a page title in shtml2html.py.
The '[<a name="top">]*' part matched any of those characters, not the
exact phrase as was presumably intended. Fix that by matching it, and
naming the group we are actually searching for as 'title'.

Fixes nss_slurm.html and pam_slurm_adopt.html having incorrect titles of
"ss_slurm" and "_slurm_adopt" respectively.