Clarify slurmdbd message comment when there's no change in data.
When some operation requests for slurmdbd returned SLURM_NO_CHANGE_IN_DATA
the comment sent back to the user was "Request didn't affect anything".

Sometimes the user only has privileges to change a subset of the things
that privileged users can change for the same operation. For these cases,
the logic doesn't directly return ESLURM_ACCESS_DENIED but instead proceeds
with a scrubbed request limited to what the user should be able to change.

This commit extends the message comment to clarify this situation.

Bug 6248.