Remove belt and suspenders error checking for sacctmgr archive load.
This was a belt and suspenders approach since calling 'open' in the mysql
code directly will produce an error code we can use to inform what is
wrong with the file given.  The file checking in sacctmgr incorrectly
assumes sacctmgr is called from the same node the slurmdbd is running on.

Bug 7244