Fix issues with --gpu-bind while using cgroups
CUDA_VISIBLE_DEVICES was not being set to the correct GPU indexes when
cgroups were being used. These issues were exhibited with at least the
map_gpu and mask_gpu binding options.

The issue was that usable_gres is a bitmask of GRESs in the step's
cgroup, but bit_test() was looking at bit i, which is the index of the
global gres_list (not constrained by cgroups).

Bug 7509