Add spank options to cache in remote callback.
When spank option callbacks are called, the options are added to a cache
in memory so that spank_option_getopt() can retrieve the options when
called. However, this was only happening when callbacks were called from
the local context, so we make sure that the options are added to the
cache when the callbacks are called from the remote context as well.

Bug 7065.