Make --batch requests wait for all nodes to boot before launching
When --batch=<feature> is used, the batch_host isn't chosen until the
job is being launched -- because the features could be different on boot
(e.g. KNL nodes). Thus if the job is allocated nodes that need to be
booted, it needs to wait till they are all booted so it can make a
decision at launch time.

Bug 7445

Signed-off-by: Dominik Bartkiewicz <bart@schedmd.com>