Fix contention for primary with "scontrol takeover [index]" where index > 1
If a backup is acting as primary and it receives a REQUEST_CONTROL from
a lower priority backup it would relinquish control but immediately
start looking for the primary and start to contend with the lower
priority backup. With this patch the lower priority backup will instead
send a REQUEST_SHUTDOWN to higher priority primaries to avoid this
contention.

Bug 7182