Docs - Clean up GRES docs.
"NVML Library" is redundant.
Add missing env var styling.
Remove awkwardly-placed, duplicated paragraph on MPS in gres.conf man
page. That information is already in gres.html. Also, saying that MPS
is required to be specified in gres.conf is wrong.
Add a reference to gres.html for a more in-depth explanation of GPU
device IDs with NVML.

Bug 7517