Fix scancel --full for proctrack/cgroups
Now the signaling of the batch step and the handeling of the flags is totally
handled in _kill_all_active_steps() in slurmd, and _handle_signal_container()
in stepd to ensure that:
- if KILL_JOB_BATCH then only batch container is signaled
- if KILL_FULL_JOB then batch script and its children are also signaled
- if both of the above then only the batch script and its children are signaled

We do not relay anymore on proctrack_g_signal() to handle the batch step
signaling anymore, therefore it works the same for all proctrack plugins.

This commit also includes minor related fixes in other code handling such
signaling flags, and documentation improvement.

Bug 7282