Rewrite test6.15: scancel signaling
Previous version relayed in the command "time".
It sent STOP and CONT signals and counted time elapsed.
The problem is that defuct childern of stopped parents are not fully
killed, and "time" keep counting time until the actual parent
continue. So, the values were wrong.

The new version uses signal handlers in a .prog and trap in the
shell to print and check if a signal is received.

NOTE: cgroups has different signaling than linuxproc and pgid.

Bug 7282