Suspend nodes after being idle or down for SuspendTime
Previously it was only after being idle. The problem was that if the
node was downed after a job ran on the node for more than SuspendTime
the node would be suspended quickly. Now it waits SuspendTime after
being idle or down (i.e. since no jobs on the node).

Bug 6774

Signed-off-by: Danny Auble <da@schedmd.com>