Fix abort initializing a configuration without acct_gather.conf.
Previous logic would only call s_p_hashtbl_create() to create the hashtable
when the file acct_gather.conf could be successfully stat()'d. This lead to
a subsequent attempt to pack the non-created hashtable into a buffer which
triggered the abort.

This makes it so the hashtable is uncondtionally created no matter if the
file is missing.

Bug 7893.