Fix cpu-bind=closest for case CPUs==Cores and ThreadPerCore!=1
gres_plugin_node_config_load in this case has to be called with the
number of CPUs defining size of the bitmap. If CPUs==Cores and
ThreadPerCore!=1 core_ids are larger than configured CPUs.

This is done every other time gres_plugin_node_config_load() is called,
this one was just missed.

Bug 7907.