Ensure x11 is setup before launching a job step
srun waits for the prolog to finish before launching a job step.
In _is_prolog_finished(), slurmctld checks the state reason:

	if (job_ptr) {
		is_running = (job_ptr->state_reason != WAIT_PROLOG);
	}

But if the job is updated during the job prolog, then _update_job() will
change the state_reason, and then slurmctld will tell srun that the
prolog is completed even if it isn't. If srun launches a job step before
the extern sets up x11, then the job step won't have x11 information. To
fix this, don't change state_reason in _update_job() if it equals
WAIT_PROLOG.

Bug 7525