Fix distributing steps on idle nodes within a job.
(1) In _pick_step_nodes(), we get a bitmap (nodes_idle) of "idle" nodes
in the job - in other words, nodes that don't have any steps running
on them.

(2) A little farther down, we check if nodes_idle has any nodes, and
try to allocate resources from the idle nodes for the step.

(3) Then we check if there are any "available" nodes in the job (nodes
that have some resources free but aren't completely idle), and try
to allocate resources from them. We will fill up a node before
allocating from the next one.

The problem: when we check if nodes are idle[1], we check which nodes
are being used by all running steps, *including* the extern and batch
steps. This means that with PrologFlags=contain (which enables the
extern step), there will never be any "idle" nodes for a job because
the extern step runs on all the nodes. We will skip section (2) and go
straight to section (3). Steps will still run, but where the steps are
allocated is different than without the extern or batch steps.

The buggy behavior is this: We pack all the steps onto the first node,
then we pack all the steps onto the second node, and so on.

The fixed behavior is the same as without a batch step or extern steps:
we run new steps on idle nodes first, then we pack subsequent steps on
each node until it is filled (only moving on to the next node once the
current node is filled or doesn't have enough resources for the step
request).

Bug 8020.