Remove REQUEST_COMPLETE_BATCH_JOB handling.
On further reflection, there is no way to receive this in 20.02,
and is safe to remove now. This is because the select/serial plugin
has been removed, and a 18.08 slurmd running select/serial will not
be able to communicate with a 20.02 cluster anyways due to the
mismatch in select plugins.