Fix regression on update from older versions with DefMemPerCPU
In 19.05 JOB_MEM_SET flag was added along with a conditional check on
this flag that changed the pn_min_memory when validating job limits.
This caused that after an upgrade, PD jobs in earlier versions didn't
have this flag and the memory was incorrectly set when their limits were
checked before starting. The patch here addresses this issue adding this
flag to jobs from an older protocol version when loading the state
files.

Bug 8011