Change the delay between successive 'sbatch --wait' queries to reduce number of RPCs.
The linear backoff up to 10s at most results in high number of
REQUEST_JOB_INFO_SINGLE RPCs being generated over time - 6 per minute
in the steady state.

With this commit we start with the same 2s delay - if the job submit
to end time is low we still return quicky - but if the job is long we
exponentially backoff up to 32 seconds to drop the steady state load
to ~2 RPCs per minute.

This is limited to a maximum backoff of (MinJobAge / 2) as well to ensure
a job does not complete, then get purged from slurmctld, all before
sbatch has had a chance to catch that transition.

Bug 8103.