Add fed_mgr_is_origin_job_id()
We need a way to tell if a job id comes from our cluster.
We already have the function fed_mgr_is_origin_job(), but
that requires a pointer to struct job_record, and we can't
get a job_ptr if the job isn't on our cluster. We need this
information when finding out if a job's dependency is on
the local cluster or a remote cluster.

Bug 6068