Break up update_job_dependency into 3 functions.
When we add logic to handle remote dependencies, the indentation in
update_job_dependency() would be too far to the right (past 80
characters), so we need to add in these helper functions. This also
significantly reduces the size of update_job_dependency() to make it
much easier to read.

Bug 6068