Detect a dependency on a remote job id.
Add logic to figure out if a dependent job id is on a remote cluster
or our local cluster. If the dependent job id is on a remote cluster
then add that job id to the dependency list. We can't test remote
dependencies for circular dependencies, so skip remote dependencies
when testing for circular dependencies. Properly print out remote
job dependencies. Hold a job that has a remote dependency so that it
doesn't start right away.

Bug 6068.