Properly detect if a reservation has really changed.
Decouple the fact that at least any of the relevant
slurmdb_reservation_rec_t members are set with the fact that any of
these actually changed with respect to the old reservation.

At this point, we could code a check so that if change variable is set
to true then place an early return before calling
acct_storage_g_modify_reservation(), avoiding this way sending a
DBD_MODIFY_RESV msg if nothing changed.

But even if nothing changed, we need to send the RPC for the edge-case
where someone just turned on accounting, and so a first record would
need to be created in the database. This implies deferring the update or
insert decision to as_mysql_modify_resv(). And this also implies that
the problem isn't really sending unneeded RPC's but creating unneeded
new database records.

Bug 5919.