Always fill in the members with the new reservation values.
We were already doing so no matter if anything changed, except for the
flags member where if nothing changed we were setting the value to
NO_VAL. Set flags also to the newest reservation value and update the
logic in as_mysql_modify_resv() in subsequent commits.

As mentioned in the previous commit, we need to send the values anyway
just in case accounting was just turned on.

Bug 5919.