Testsuite - Fix sleep in test17.12
Increase the sleep in job to 5s was required to make test reliable. It
doesn't result in longer execution since we're waiting for job to get
into RUNNING state and then we're canceling it. Short sleep can end-up
with overlook of RUNNING state and false negative result - 'Job is DONE
but expected RUNNING'.

Bug 7624.