Testsuite - Fix test job name to avoid false positives on test12.10
Make the job name searched for by the test include `test${test_id}-`
so it does not mistake a job id as the job name.
Replace the random number with a timestamp to guarantee that there
are no repeats.
Add `-p` to the sacct arguments to accommodate the increased length
of the job name and to guarantee that the output never gets truncated.

Bug 7212