Detect if only remote dependencies remain.
Add DEPEND_REMOTE as a new return value for test_job_dependency() and
add logic to return DEPEND_REMOTE if only remote dependencies remain. We
will use this later to tell if all local dependencies are satisfied on a
sibling so the sibling can inform the origin cluster when there are no
more pending dependencies.

Bug 6068