Testsuite - Refactor test1.35 avoiding random sleeps and wasted time
Test was running squeue inside the batch job and checking the output
file.
The refactor simplifies the batch job moving squeue out of it, no
random sleep values are necessary and uses wait_for to check that
the expected steps are on squeue (outside the batch job).
We don't need to wait for the job to end anymore, but we can cancel
it.
Adding also a cleanup as a best practice.

Bug 8231

Signed-off-by: Marcin Stolarek <cinek@schedmd.com>