Keep track of dependency state
Rather than deleting job dependencies as they are fulfilled, keep track
of the state of each dependency (not fulfilled, fulfilled, failed).
Don't delete dependencies unless they are OR satisfied.

We will use this in a future commit to send fulfilled/failed
dependencies from a sibling slurmctld back to the origin slurmctld so
the origin knows how to update the job's dependency list. If
dependencies are OR satisfied (and thus the dependency list is empty),
then the sibling will tell the origin that the dependencies were OR
satisfied.

Bug 6068