Fix DependencyNeverSatisfied for OR'd dependencies
Previous behavior:
If the last dependency in the list failed, then the job would get marked
as DependencyNeverSatisfied. If the sbatch flag
--kill-on-invalid-dep=yes or the SchedulerParameters option
kill_invalid_depend was used, then the job would be killed. These things
happened even if there were other dependencies that might be fulfilled.

With this commit:
If there are any unfulfilled dependencies in an OR'd dependency list,
the job will no longer be marked as DependencyNeverSatisfied.

Found while working on bug 6068. First reported in bug 5851 comment 1.

Bug 5851