Improve performance of xstrndup().
Found while working to improve sbatch script parsing speeds.

It turns out that the call to strlcpy() was the culprit of 97% or more
of the runtime, specifically the second part that iterated through the
source string to find the end. That result was only used for the return
value, which for xstrndup(), the function that sbatch was using, is
ignored.

In other words, sbatch used xstrndup(), which called strlcpy() but
ignored the return value, and it was in the calculation of the return
value of strlcpy() that 97% or more of the time was taken.

I changed xstrndup to use memcpy() instead. This made sbatch much
faster, going from more than 2 hours to process a 75 MB script to 1.5
seconds.

Bug 7812.