Test remote job dependencies on siblings.
This sets up the architecture for testing remote job dependencies when
they're on the sibling clusters. We add a list of job with remote
dependencies, add a function to iterate through that list and call
test_job_dependency() on each of them, and call that from
purge_old_job() and from a new thread so that remote job dependencies
will be tested regularly. It works, but we aren't doing anything with
the results yet (we still need to tell the origin cluster about updates
to jobs with remote dependencies).

Bug 6068