Setup for updating origin job dependencies.
Improve error handling. Move the internal while() loop in
_orig_dep_update_thread() to a new function, because we'll be adding a
lot of stuff and indentation might go too far in. Add todo's in
comments. Move the lock_slurmctld to be around the whole while() loop
instead of inside of it, because later we'll need it to be locked
outside the loop when we handle the TODO's.

Bug 6068