Move _build_path() after SPANK and other plugins have run.
If, e.g., a SPANK plugin needs to manipulate the filesystem namespaces
we want that to complete first so that we can discover an executable
within them.

Reword the comment as part of relocating this call as it was backwards.
MPMD, before the previous commit, did handle PATH resolution locally
within srun, whereas normal task launch has deferred to the slurmstepd
for a while.

Bug 7257.