Send remote dependency updates as they change.
Instead of waiting for the entire dependency for the job to be fulfilled
or failed, send an update of the job dependency to the origin as each
dependency is fulfilled. That way the job dependency has the most recent
information, and querying the job with scontrol show job will show the
updated dependency.

Bug 6068