Save an RPC for testing remote dependencies.
When a sibling receives remote dependencies for a job, only add the job
to the list to test if there was at least one dependency local to the
sibling. If there aren't any dependencies local to the sibling, then
there's no point in adding the job to that list, testing the
dependencies, and sending the origin an RPC saying that all dependencies
are cleared.

Bug 6068