Make multi-cluster singleton dependencies work.
Singleton dependencies now work for federated clusters.

Limitation:
As soon as the singleton dependency is satisfied on a sibling,
that cluster will tell the origin that the singleton dependency has been
satisfied on that cluster and won't check anymore for that singleton,
even if another job from the same user with the same name starts on that
sibling.

Bug 6068