Don't add the singleton dependent job to the list of jobs to check.
This is a minor optimization. When calling _build_user_job_list to
create a list of jobs of the same user and job name, we added the
dependent job, too. This didn't affect the check in
test_job_dependency() because we checked if the job_id was before our
own, but there's no reason to add the job because it doesn't check
itself.

Bug 6068