Remove dependencies on siblings when job dependency is fulfilled
When a job dependency is OR'd, when any one of the dependencies is
fulfilled, the job's dependency is satisfied. When the job has
dependencies on multiple sibling clusters and a dependency any one
cluster is satisfied, this commit removes dependencies on all siblings
so that siblings don't keep testing unfulfilled dependencies for a job
whose dependency was already fulfilled.

Bug 6068