Remove unnecessary check against LINE_MAX in _register_oom_notifications().
This was inherited from the cgroup_event_listener.c code this was based
on originally, but was only needed there as the line buffer was a fixed
length. Since we're using an xstring here this is not a concern.

And, as we're printing just two integers here, this line could never
possibly exceed any POSIX compliant value for LINE_MAX, which is
required to be >= (_POSIX2_LINE_MAX == 2048).

This also removes the last place using the return value from xstrfmtcat().