Add cosmetic batch step to correct component in a hetjob.
Before this the batch step was always added to the last component
instead of the first internally in the slurmctld.

The batch step was ran correctly, just not accounted for correctly.

This fixes the error when a batch step completes...

error: _slurm_rpc_complete_batch_script: Could not find batch step for
 JobId=xxxx+0(xxxx), this should never happen


Signed-off-by: Marshall Garey <marshall@schedmd.com>