Track dependency state in state save location.
The dependency string is saved in StateSaveLocation. But since we are
keeping invalid dependencies in the string, we need to save the state of
the dependency, which we do in this commit. The previous behavior was
an invalid dependency would get restored on slurmctld startup, but we
wouldn't know what state it was in. So, it would get tested by
test_job_dependency() and be considered fulfileld because we couldn't
find the job id. We can avoid this by saving the state of each
dependency as part of the dependency string. This allows us to keep
tracking invalid dependencies in the dependency string.

Bug 6068