Move dependency options in slurm.conf to DependencyParameters.
Move kill_invalid_depend from SchedulerParameters to
DependencyParameters. Deprecate it properly - we still parse it in
SchedulerParameters, but print a warning that it is deprecated and
should be used in DependencyParameters. Also add to RELEASE_NOTES that
it is deprecated.

Move disable_multicluster_singleton from SlurmctldParameters to
DependencyParameters and rename it to disable_remote_singleton.

Bug 6068