Correctly set remote flag on restoring remote dependencies.
Ensure that update_job_dependency() is called for each job in
remote_dep_job_list after fed_mgr_cluster_rec is set so that we can
correctly set the remote depend flag for each dependency. This matters
when we're restoring job dependencies from the state file on slurmctld
startup.

Bug 6068