Send singleton dependencies to clusters added to a federation.
When a cluster is added to a federation, send singleton dependencies
from clusters already in the federation to it so it can test the
singleton dependencies. If we don't do this, singleton dependencies will
remain unfulfilled forever until that cluster is removed from the
federation because singleton dependencies are only fulfilled when every
cluster currently in the federation have told the origin that it's
fulfilled.

Bug 6068