Immediately fail dependencies on clusters that leave a federation.
When a cluster leaves a federation, immediately test all job dependencies
so that dependencies on jobs in the cluster that left will fail.
Dependencies on jobs in clusters that aren't part of the federation should
fail, so we just do this right away instead of waiting for the scheduler
to call job_independent() -> test_job_dependency(). Also, the sibling
already clears out remote_dep_job_list when it leaves the federation.

Bug 6068