Overhaul fetch_config() to handle issues with stray process state.
This is not ideal, but provides a workaround to stray process state
meaning that a series of operation like:
	slurm_conf_init(A) -> slurm_conf_destroy() -> slurm_conf_init(B)
does not lead to the same process state that
	slurm_conf_init(B)
would.

This is due to a number of local cached values strewn throughout the source
code. For example, (as of this commit), _slurm_connect() in
slurm_protocol_socket.c uses a timeout value that is loaded once with
slurm_get_tcp_timeout() and will never change again.

Some places do handle reconfiguration correctly, e.g., _schedule() in
job_scheduler.c checks slurmctld_conf.last_update to see if the cached
values should be overhauled.

However, this does not help code in src/common, which must check
slurmdbd_conf.last_update iff running in the slurmdbd instead.
(And slurmdbd_conf.last_update is never updated at present.)

So... use this fork()'d process to fetch the config, and send them
back up to us over a pipe. That process is thus free to dirty its local
state without affecting the main process.