Avoid slurmctld falling into "configless" operation.
This would only have happened if (a) no -f option was specified,
(b) the SLURM_CONF envvar was not set, and (c) default_slurm_config_file
was not available.

But if it does happen it won't work, since slurmctld will need to be
started to respond to the RPC request.