slurmd.service - Disable ConditionPathExists to permit configless operation.
In configless operation this file is intentionally omitted, thus blocking
the slurmd service from starting until it exists is counterproductive.

If you are running and require this config, the "configless" mode of
operation will always fall back to the configuration file as a last
resort, and spend up to 60 seconds trying to load it.