Testsuite - Fix intermittent failures in test 38.4
When there is too much output going through expect's output parsing,
random errors can happen. So the fix is to simply print env directly
to a log file rather than parsing and printing it.
Also, instead of catting the entire output file, grep it so that
expect is only dealing with a few lines.

Bug 8418