Send GRES context from slurmd to stepd to avoid rebuilding it
This avoids slow context rebuilding code in gres_plugin_init(). Since
we are already sending over config_flags, we might as well just send
over as much of the GRES context as we can.

Instead of writing it every time for a step we build a buffer once
and then send that every time as done elsewhere.

Bug 7985

Co-authored-by: Danny Auble <da@schedmd.com>