Send GRES config to stepd and remove stepd's node_config_load()
Pass the merged slurm.conf/gres.conf gres_conf_list and
autodetect_types to the stepd from the slurmd, and remove the call to
node_config_load() in the stepd. Doing this can drastically improve
the stepd startup time when GPU binding is used with NVML, since NVML
initialization can take on the order of seconds.

Note that setting the GPU frequency will still require the stepd to
initialize NVML in order to access the NVML API, but because
we pass autodetect_types to the stepd as well, we do not need to
re-parse gres.conf with node_config_load() to look for AutoDetect.

Bug 7985

Co-authored-by: Danny Auble <da@schedmd.com>