Correct parsing of depend_time.
If a dependency didn't have a time limit but then the one after it does,
the time limit was attached to the first dependency and the second
dependency was dropped. An example:

$ squeue
    JOBID PARTITION     NAME     USER ST       TIME  NODES NODELIST(REASON)
134218254     debug     wrap marshall  R       1:04      1 d2_1
134218253     debug     wrap marshall  R       6:36      1 d2_1

$ sbatch --depend=afterok:134218253,after:134218254+1 -Dtmp \
		  --wrap="srun whereami 20"
Submitted batch job 67109108

$ scontrol show job |egrep -i "name|depend"
JobId=67109108 JobName=wrap
   JobState=PENDING Reason=Dependency \
	    Dependency=afterok:134218253+1(unfulfilled)

The parsing logic did a strchr to find the '+' character, thus skipping
over the entirety of ",after:134218254" and attaching the time to
"afterok:134218253"

Replacing strchr with tmp[0] == '+' fixes this.

Bug 8456